Add files via upload
In this course project, a Telegram quest bot called "Bandero-Husaka" will be
developed using the Java programming language. The application works in
the format of a quest, where the user is presented with a set of tasks and upon completing
them, receives a reward in the form of coins. Additionally, the user is provided with an
NPC, Bandero-Husak, whom they can level up as a biological weapon by giving him
titles such as "Ordinary Husak," "Biogusak," "Bandero-Stager," and "Bandero-Husak."
At the end of the game, the user can purchase a Javelin and go to silence the monster.